Rewrite algorithm to be more constant time
The previous approach would figure out the reverse dependencies of a
package by iterating over all packages for finding the reverse
dependencies of a package. Now we create one hashmap with the package to
reverse dependencies mapping which reduces the amount of work we have to
do and makes the program more or less constant time.