Handle provided packages of packages in pkgnames
Packages in our repository have makedepends to a "provide" for example
irccat has a makedepend on go-pie which is provided by go. As we do not
want to have provides in our rebuild list we can not simply only add it
as a new package in pkgnames, the algorithm now "resolves" provided
pkgnames to the repository package.