Sort the reverse dependencies so output is predictable
We build a reverse dependency map which contains a HashSet of reverse
dependencies which are not sorted. This leads to the graph being
generated different on multiple runs which in return makes the output
not predictable. While not being a "rebuild order" issue, it might lead
to confusion.

Merging the pkgnames into one DFS graph also resolves the issue of
multiple inputs not being merged in output.

Fixes: #11, #12