lib: Implement '--no-reverse-depends' that creates a dependency chain from stdin
When building closed loops of packages you sometime don't want the
entire chain to be resolved to just figure out the order of 5 packages.
This implement `--no-reverse-depends` which resolves the chain and
ensures only the packages from stdin are included in the output chain.

This is useful for rebuild lists or when building tight loops of
packages.

    $ arch-rebuild-order --no-reverse-depends cmake meson jsoncpp
    meson jsoncpp cmake
    $ arch-rebuild-order --no-reverse-depends  jsoncpp meson cmake
    meson jsoncpp cmake

Signed-off-by: Morten Linderud <morten@linderud.pw>
Co-Authored-By: Jelle van der Waa <jelle@archlinux.org>